requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
Pillow==10.0.0
instagrapi==2.0.0
python-dateutil==2.8.2
schedule==1.2.0